+++ android/gitudy/data/src/main/java/com/takseha/data/dto/auth/login/Role.kt
@@ -1,4 +1,4 @@
-package com.takseha.data.dto.login.login
+package com.takseha.data.dto.auth.login
 
 enum class Role {
     ADMIN, UNAUTH, USER, WITHDRAW
