+++ android/gitudy/data/src/main/java/com/takseha/data/dto/auth/register/RegisterResponse.kt
@@ -1,22 +1,20 @@
-package com.takseha.data.dto.login.login
+package com.takseha.data.dto.auth.register
 
 
 import com.google.gson.annotations.SerializedName
 
-data class LoginResponse(
+data class RegisterResponse(
     @SerializedName("res_code")
     val resCode: Int,
     @SerializedName("res_msg")
     val resMsg: String,
     @SerializedName("res_obj")
-    val tokenInfo: LoginTokenInfo
+    val tokenInfo: RegisterTokenInfo
 )
 
-data class LoginTokenInfo(
+data class RegisterTokenInfo(
     @SerializedName("access_token")
     val accessToken: String,
     @SerializedName("refresh_token")
-    val refreshToken: String,
-    @SerializedName("role")
-    val role: String
+    val refreshToken: String
 )
